requirements.txt
streamlit
python-chess
numpy
pandas

#newest packages as of october 19, 2024 i suppose.